refactor: delete src/utils/file-operations.ts; inline single-use helpers
Each helper had exactly one production caller (generateProject.ts) and
most added no semantic value over the stdlib call they wrapped:

- ensureDir: mkdirSync with {recursive: true} is already idempotent;
  inlined.
- pathExists: one-line rename of existsSync; inlined.
- getIsTemplateFile: pre-check looped the same keys replacePlaceholders
  was about to loop. replaceAll is a no-op when the key is missing, so
  the guard only saved a string copy. Dropped.
- getFilesRecursive: one-line wrap around Bun.Glob().scanSync(); the
  existsSync early-return was dead in practice. Inlined into
  DEFAULT_SOURCE_READER.list.
- replacePlaceholders: only export with real semantic content. Moved
  next to planTemplate in generateProject.ts (its single consumer) and
  retained as an export for tests.

Tests: deleted tests/file-operations.test.ts (was covering thin fs
wrappers); folded the replacePlaceholders cases into
tests/plan-template.test.ts. The integration test now uses a local
listFilesRecursive helper.
